Created separate header/navbar component, implemented better styling for mobile
