+++ backend/.env
@@ -1,0 +1,1 @@
+MONGODB_URI=mongodb+srv://gabeaz03:jKsBkJMciukxRj4F@testing.4lktkzy.mongodb.net/?retryWrites=true&w=majority&appName=Testing
